+++ release/src-rt-6.x.4708/router/libcurl/tests/libtest/lib3010.c
@@ -39,8 +39,7 @@ static CURLcode test_lib3010(const char *URL)
     if(result) {
       curl_mfprintf(stderr,
                     "%s:%d curl_easy_perform() failed with code %d (%s)\n",
-                    __FILE__, __LINE__, (int)result,
-                    curl_easy_strerror(result));
+                    __FILE__, __LINE__, result, curl_easy_strerror(result));
       goto test_cleanup;
     }
     curl_easy_getinfo(curl, CURLINFO_REDIRECT_URL, &follow_url);
@@ -51,8 +50,7 @@ static CURLcode test_lib3010(const char *URL)
     if(result) {
       curl_mfprintf(stderr,
                     "%s:%d curl_easy_perform() failed with code %d (%s)\n",
-                    __FILE__, __LINE__, (int)result,
-                    curl_easy_strerror(result));
+                    __FILE__, __LINE__, result, curl_easy_strerror(result));
       goto test_cleanup;
     }
